Separate WriteBufferManager stats ownership
Move process-wide stats and watchdog lifecycle out of DBSettings, keep the stats API module-scoped, and prove process.exit teardown for both global-listener construction orders.
