+++ binding.gyp
@@ -74,6 +74,7 @@
 				'src/binding/database/db_descriptor.cpp',
 				'src/binding/database/db_handle.cpp',
 				'src/binding/database/db_registry.cpp',
+				'src/binding/database/db_stats.cpp',
 				'src/binding/database/db_settings.cpp',
 				'src/binding/iterator/db_iterator.cpp',
 				'src/binding/iterator/db_iterator_handle.cpp',
