+++ src/binding/database/db_settings.cpp
@@ -1,10 +1,7 @@
 #include "database/db_settings.h"
-#include <cstdio>
 #include <random>
+#include "database/db_stats.h"
 #include "napi/macros.h"
-#include "core/platform.h"
-#include "database/db_registry.h"
-#include "napi/global_events.h"
 #include "napi/helpers.h"
 #include "napi/async.h"
 #include "rocksdb/advanced_cache.h"
@@ -18,44 +15,6 @@ uint64_t generateSeed() {
 	uint64_t hi = static_cast<uint64_t>(rd());
 	uint64_t lo = static_cast<uint64_t>(rd());
 	return (hi << 32) | lo;
-}
-
-// Resolved once per process: the watchdog runs on its own thread, and ::getenv
-// there is not safe against a concurrent process.env write on any JS thread
-// (same reasoning as ROCKSDB_JS_PARK_TIMEOUT_MS).
-uint64_t writeBufferManagerStallWarnMs() {
-	static const uint64_t value = [] {
-		const char* raw = ::getenv("ROCKSDB_JS_WBM_STALL_WARN_MS");
-		bool rejected = false;
-		uint64_t resolved = resolveWbmStallWarnMs(raw, &rejected);
-		if (rejected) {
-			::fprintf(stderr,
-				"[rocksdb-js] ignoring ROCKSDB_JS_WBM_STALL_WARN_MS=\"%s\" (not an integer in "
-				"[0, %llu] ms); using %llu\n",
-				raw, static_cast<unsigned long long>(WBM_STALL_WARN_MS_MAX),
-				static_cast<unsigned long long>(resolved));
-		}
-		return resolved;
-	}();
-	return value;
-}
-
-napi_status setNumberProperty(napi_env env, napi_value target, const char* key, uint64_t value) {
-	napi_value jsValue;
-	napi_status status = ::napi_create_double(env, static_cast<double>(value), &jsValue);
-	if (status != napi_ok) {
-		return status;
-	}
-	return ::napi_set_named_property(env, target, key, jsValue);
-}
-
-napi_status setBoolProperty(napi_env env, napi_value target, const char* key, bool value) {
-	napi_value jsValue;
-	napi_status status = ::napi_get_boolean(env, value, &jsValue);
-	if (status != napi_ok) {
-		return status;
-	}
-	return ::napi_set_named_property(env, target, key, jsValue);
 }
 
 } // namespace
@@ -133,182 +92,12 @@ std::shared_ptr<rocksdb::WriteBufferManager> DBSettings::getWriteBufferManager()
 			cache,
 			writeBufferManagerAllowStall.load(std::memory_order_relaxed)
 		);
-		writeBufferManagerPtr.store(writeBufferManager.get(), std::memory_order_release);
+		DBStats::getInstance().publishWriteBufferManager(writeBufferManager.get());
 	}
 	if (writeBufferManagerAllowStall.load(std::memory_order_relaxed)) {
-		this->ensureWriteBufferManagerWatchdog();
+		DBStats::getInstance().ensureWriteBufferManagerWatchdog();
 	}
 	return writeBufferManager;
-}
-
-/**
- * Starts the stall watchdog if it is not already running. Only a stalling manager
- * can stall (`WriteBufferManager::ShouldStall` short-circuits on `allow_stall`),
- * so nothing is started otherwise.
- *
- * Thread creation is allowed to fail without taking the caller down with it: this
- * is an observability feature, and turning thread exhaustion into a failed
- * `config()` or database open would be strictly worse than losing the warn line.
- * The failure is visible as `watchdogRunning: false`.
- */
-void DBSettings::ensureWriteBufferManagerWatchdog() {
-	if (writeBufferManagerStallWarnMs() == 0) {
-		return;
-	}
-	std::lock_guard<std::mutex> lock(this->watchdogMutex);
-	// A stop that has been requested but not yet joined is left to its joiner:
-	// this path runs under databasesMutex (DBRegistry::OpenDB holds it across
-	// DBDescriptor::open), and the watchdog takes databasesMutex to collect its
-	// inventory, so joining here could deadlock against a sample in flight.
-	if (this->watchdogStarted) {
-		return;
-	}
-	this->watchdogStopRequested = false;
-	this->writeBufferManagerWatchdogStopping.store(false, std::memory_order_relaxed);
-	const uint64_t generation = ++this->watchdogGeneration;
-	try {
-		this->watchdogThread =
-			std::thread([this, generation]() { this->runWriteBufferManagerWatchdog(generation); });
-		this->watchdogStarted = true;
-		// Set here, not on the new thread: a caller that enables stalling and reads
-		// back immediately must not be told the watchdog is absent.
-		this->writeBufferManagerWatchdogRunning.store(true, std::memory_order_relaxed);
-		if (!this->watchdogAtexitRegistered) {
-			// Registered while the thread starts, so on the process.exit() path
-			// (which skips the module cleanup hook) the join runs ahead of every
-			// static destructor registered earlier — DBRegistry's among them, since
-			// its instance is a class static registered before main while this
-			// object is first touched by config(). ~DBSettings stays as the
-			// backstop; both calls are idempotent.
-			this->watchdogAtexitRegistered =
-				::atexit([]() { DBSettings::getInstance().joinWriteBufferManagerWatchdog(); }) == 0;
-		}
-	} catch (...) {
-		this->watchdogStarted = false;
-		this->writeBufferManagerWatchdogRunning.store(false, std::memory_order_relaxed);
-	}
-}
-
-void DBSettings::requestWriteBufferManagerWatchdogStop() {
-	{
-		std::lock_guard<std::mutex> lock(this->watchdogMutex);
-		this->watchdogStopRequested = true;
-		this->writeBufferManagerWatchdogStopping.store(true, std::memory_order_relaxed);
-	}
-	this->watchdogCv.notify_all();
-}
-
-void DBSettings::joinWriteBufferManagerWatchdog() {
-	std::thread toJoin;
-	{
-		std::lock_guard<std::mutex> lock(this->watchdogMutex);
-		this->watchdogStopRequested = true;
-		this->writeBufferManagerWatchdogStopping.store(true, std::memory_order_relaxed);
-		if (this->watchdogStarted) {
-			toJoin = std::move(this->watchdogThread);
-			this->watchdogStarted = false;
-		}
-	}
-	this->watchdogCv.notify_all();
-	if (toJoin.joinable()) {
-		toJoin.join();
-	}
-}
-
-void DBSettings::runWriteBufferManagerWatchdog(uint64_t generation) {
-	setThreadName("rocksdb-wbm-watchdog");
-	const uint64_t thresholdMs = writeBufferManagerStallWarnMs();
-	WbmStallWatchdogState state;
-	std::unique_lock<std::mutex> lock(this->watchdogMutex);
-	auto retired = [&] {
-		return this->watchdogStopRequested || this->watchdogGeneration != generation;
-	};
-	while (!retired()) {
-		this->watchdogCv.wait_for(lock, std::chrono::milliseconds(WBM_STALL_SAMPLE_INTERVAL_MS));
-		if (retired()) {
-			break;
-		}
-		lock.unlock();
-		try {
-			this->sampleWriteBufferManagerStall(state, thresholdMs);
-		} catch (...) {
-			// A report that could not be built (allocation, formatting) must not
-			// take the thread down: the next sample retries.
-		}
-		lock.lock();
-	}
-	if (this->watchdogGeneration == generation) {
-		this->writeBufferManagerStallActiveMs.store(0, std::memory_order_relaxed);
-		this->writeBufferManagerWatchdogRunning.store(false, std::memory_order_relaxed);
-	}
-}
-
-void DBSettings::sampleWriteBufferManagerStall(WbmStallWatchdogState& state, uint64_t thresholdMs) {
-	rocksdb::WriteBufferManager* wbm = this->getWriteBufferManagerRaw();
-	if (wbm == nullptr) {
-		return;
-	}
-	WbmStallWatchdogState::Sample sample = state.onSample(
-		wbm->IsStallActive(), WbmStallWatchdogState::Clock::now(), thresholdMs
-	);
-	this->writeBufferManagerStallActiveMs.store(sample.stallActiveMs, std::memory_order_relaxed);
-	if (!sample.reportNow ||
-		this->writeBufferManagerWatchdogStopping.load(std::memory_order_relaxed)) {
-		return;
-	}
-
-	WriteBufferManagerStallReport report;
-	report.stallActiveMs = sample.stallActiveMs;
-	report.bufferSize = wbm->buffer_size();
-	report.memoryUsage = wbm->memory_usage();
-	report.mutableMemoryUsage = wbm->mutable_memtable_memory_usage();
-	report.allowStall = this->writeBufferManagerAllowStall.load(std::memory_order_relaxed);
-	report.costToCache = this->writeBufferManagerCostToCache.load(std::memory_order_relaxed);
-	report.inventoryAvailable = DBRegistry::CollectWriteBufferManagerInventory(
-		wbm, report.columnFamilies, report.maxWriteBufferSizeToMaintain
-	);
-
-	std::string line = formatWriteBufferManagerStallReport(report);
-	// stderr does not depend on an application having registered a listener; the
-	// event is the one an application can route. The episode is retired once
-	// either has carried it — gating on stderr alone would re-report every second
-	// for the whole stall when fd 2 is closed but a listener is attached.
-	const bool wroteToStderr = ::fprintf(stderr, "%s\n", line.c_str()) >= 0;
-	if (wroteToStderr) {
-		::fflush(stderr);
-	}
-	const bool emitted = emitGlobalEvent("log.warn", ListenerData::fromStrings({ line }));
-	if (wroteToStderr || emitted) {
-		state.markReported();
-	}
-}
-
-WriteBufferManagerStats DBSettings::getWriteBufferManagerStats(bool includeColumnFamilies) {
-	WriteBufferManagerStats stats;
-	stats.allowStall = this->writeBufferManagerAllowStall.load(std::memory_order_relaxed);
-	stats.costToCache = this->writeBufferManagerCostToCache.load(std::memory_order_relaxed);
-	stats.watchdogRunning = this->writeBufferManagerWatchdogRunning.load(std::memory_order_relaxed);
-
-	rocksdb::WriteBufferManager* wbm = this->getWriteBufferManagerRaw();
-	if (wbm == nullptr) {
-		return stats;
-	}
-	stats.enabled = true;
-	stats.bufferSize = wbm->buffer_size();
-	stats.memoryUsage = wbm->memory_usage();
-	stats.mutableMemoryUsage = wbm->mutable_memtable_memory_usage();
-	stats.stallActive = wbm->IsStallActive();
-	stats.stallActiveMs = this->writeBufferManagerStallActiveMs.load(std::memory_order_relaxed);
-	if (includeColumnFamilies) {
-		stats.inventoryAvailable = DBRegistry::CollectWriteBufferManagerInventory(
-			wbm, stats.columnFamilies, stats.maxWriteBufferSizeToMaintain
-		);
-	}
-	return stats;
-}
-
-DBSettings::~DBSettings() {
-	this->joinWriteBufferManagerWatchdog();
 }
 
 /**
@@ -425,12 +214,8 @@ napi_value DBSettings::Config(napi_env env, napi_callback_info info) {
 		if (wbmAlreadyCreated && allowStallProvided) {
 			settings.writeBufferManager->SetAllowStall(newAllowStall);
 			if (newAllowStall) {
-				// The databases already attached can stall from here on, with no
-				// further open to start the watchdog.
-				settings.ensureWriteBufferManagerWatchdog();
+				DBStats::getInstance().ensureWriteBufferManagerWatchdog();
 			} else {
-				// Nothing can stall any more, so the thread would poll forever and
-				// report watchdogRunning against allowStall: false.
 				retireWatchdog = true;
 			}
 		}
@@ -441,7 +226,7 @@ napi_value DBSettings::Config(napi_env env, napi_callback_info info) {
 	// writeBufferManagerMutex. Same split, and the same reason, as the teardown
 	// path in binding.cpp.
 	if (retireWatchdog) {
-		settings.joinWriteBufferManagerWatchdog();
+		DBStats::getInstance().joinWriteBufferManagerWatchdog();
 		// Reconcile against the state that is actually current, not this call's
 		// own argument: another env re-enabling stalling in the unlocked window
 		// above would have found the thread still started and declined to start
@@ -449,7 +234,7 @@ napi_value DBSettings::Config(napi_env env, napi_callback_info info) {
 		std::lock_guard<std::mutex> lock(settings.writeBufferManagerMutex);
 		if (settings.writeBufferManagerAllowStall.load(std::memory_order_relaxed) &&
 			settings.writeBufferManager) {
-			settings.ensureWriteBufferManagerWatchdog();
+			DBStats::getInstance().ensureWriteBufferManagerWatchdog();
 		}
 	}
 
@@ -475,44 +260,7 @@ napi_value DBSettings::Config(napi_env env, napi_callback_info info) {
 }
 
 /**
- * The `getWriteBufferManagerStats()` JavaScript function. Process-wide: the
- * WriteBufferManager is a singleton shared by every database opened in this
- * process, including from worker threads.
- */
-napi_value DBSettings::GetWriteBufferManagerStats(napi_env env, napi_callback_info info) {
-	WriteBufferManagerStats stats = DBSettings::getInstance().getWriteBufferManagerStats(true);
-
-	napi_value result;
-	NAPI_STATUS_THROWS(::napi_create_object(env, &result));
-	NAPI_STATUS_THROWS(setNumberProperty(env, result, "bufferSize", stats.bufferSize));
-	NAPI_STATUS_THROWS(setNumberProperty(env, result, "memoryUsage", stats.memoryUsage));
-	NAPI_STATUS_THROWS(setNumberProperty(env, result, "mutableMemoryUsage", stats.mutableMemoryUsage));
-	NAPI_STATUS_THROWS(setNumberProperty(env, result, "stallActiveMs", stats.stallActiveMs));
-	NAPI_STATUS_THROWS(setNumberProperty(env, result, "columnFamilies", stats.columnFamilies));
-	NAPI_STATUS_THROWS(setBoolProperty(env, result, "enabled", stats.enabled));
-	NAPI_STATUS_THROWS(setBoolProperty(env, result, "allowStall", stats.allowStall));
-	NAPI_STATUS_THROWS(setBoolProperty(env, result, "costToCache", stats.costToCache));
-	NAPI_STATUS_THROWS(setBoolProperty(env, result, "stallActive", stats.stallActive));
-	NAPI_STATUS_THROWS(setBoolProperty(env, result, "watchdogRunning", stats.watchdogRunning));
-	NAPI_STATUS_THROWS(setBoolProperty(env, result, "inventoryAvailable", stats.inventoryAvailable));
-
-	napi_value targets;
-	NAPI_STATUS_THROWS(::napi_create_object(env, &targets));
-	for (const auto& [target, count] : stats.maxWriteBufferSizeToMaintain) {
-		napi_value countValue;
-		NAPI_STATUS_THROWS(::napi_create_int64(env, static_cast<int64_t>(count), &countValue));
-		NAPI_STATUS_THROWS(::napi_set_named_property(
-			env, targets, std::to_string(target).c_str(), countValue
-		));
-	}
-	NAPI_STATUS_THROWS(::napi_set_named_property(env, result, "maxWriteBufferSizeToMaintain", targets));
-
-	return result;
-}
-
-/**
- * Exports the `config()` and `getWriteBufferManagerStats()` functions to
- * JavaScript.
+ * Exports the `config()` function to JavaScript.
  *
  * @param env The Node.js environment.
  * @param exports The exports object.
@@ -529,19 +277,6 @@ void DBSettings::Init(napi_env env, napi_value exports) {
 	));
 
 	NAPI_STATUS_THROWS_VOID(::napi_set_named_property(env, exports, "config", configFn));
-
-	napi_value wbmStatsFn;
-	NAPI_STATUS_THROWS_VOID(::napi_create_function(
-		env,
-		"getWriteBufferManagerStats",
-		NAPI_AUTO_LENGTH,
-		DBSettings::GetWriteBufferManagerStats,
-		nullptr,
-		&wbmStatsFn
-	));
-	NAPI_STATUS_THROWS_VOID(
-		::napi_set_named_property(env, exports, "getWriteBufferManagerStats", wbmStatsFn)
-	);
 }
 
 }
